Rename light-spike crate to light-client
