+++ light-client/Cargo.toml
@@ -1,10 +1,8 @@
 [package]
-name = "light-spike"
+name = "light-client"
 version = "0.1.0"
 authors = ["Romain Ruetschi <romain@informal.systems>"]
 edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 tendermint = { path = "../tendermint" }
