+++ light-client/src/components/verifier.rs
