+++ light-client/src/contracts.rs
