+++ light-client/src/errors.rs
