+++ light-client/src/operations/voting_power.rs
