+++ light-client/src/predicates.rs
