+++ light-client/src/predicates/errors.rs
