+++ light-client/src/state.rs
