+++ light-client/src/store.rs
