+++ light-client/src/store/memory.rs
