+++ light-client/src/types.rs
